Add --boot-mode to select either nvram or startup.nsh
We can load the tests in one of 2 ways now, either via nvram (like you'd
expect a system to have) or startup.nsh. startup.nsh is more likely
if someone came to a uefi shell and typed: kernel.efi arg1 ...

in the future maybe we'll put this 'mode' into the yaml, but this
was quicker for now and allows me to excersise the path.